pyproject.toml
[tool.poetry]
name = "water-level"
version = "1.0.0"
description = ""
authors = ["pwielgor <pwielgor@camk.edu.pl>"]
readme = "README.md"
packages = [{include = "water_level"}]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = ">=3.5"
oca_nats_config = {git = "https://github.com/araucaria-project/oca_nats_config.git"}
serverish = {git = "https://github.com/AkondLab/serverish.git"}


[tool.poetry.scripts]
water_level = "water_level.read_from_arduino:main"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
